test(segcache): ABA regression test for stale CAS tokens
Deterministically reproduce the segment-recycle ABA: insert a key,
capture its token, clear() the cache (push_free prepends the freed
segment to the free queue and bumps its generation), then reinsert
the same key — it lands at the identical 44-bit location. The stale
token must now fail with Exists; with location-only tokens it falsely
succeeded. Precondition asserts verify the test really reproduces the
same-location scenario, so it fails loudly if free-queue ordering
ever changes.

Verified the test bites: with the generation bump removed, stale and
fresh tokens are identical (the false-success case).

Also widen the fuzz target's cas argument to u64.

+++ crates/segcache/fuzz/fuzz_targets/segcache.rs
@@ -178,7 +178,7 @@ fuzz_target!(|data: &[u8]| {
                         key, value, ttl
                     );
                 }
-                let _ = cache.cas(key, value, None, Duration::from_secs(ttl.into()), cas);
+                let _ = cache.cas(key, value, None, Duration::from_secs(ttl.into()), u64::from(cas));
             }
             6 => { // incr
                 if i >= data.len() {
@@ -323,7 +323,7 @@ fuzz_target!(|data: &[u8]| {
                         key, value, ttl
                     );
                 }
-                let _ = cache.cas(key, value, None, Duration::from_secs(ttl.into()), cas);
+                let _ = cache.cas(key, value, None, Duration::from_secs(ttl.into()), u64::from(cas));
             }
             _ => {
                 return;
